Componentize AssignToProductDropdown
